# lib/api-zod/src/generated/types/siteContentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteContentInput {
  /**
     * @minLength 1
     * @maxLength 60000
     */
  termsContent: string;
  /**
     * @minLength 1
     * @maxLength 60000
     */
  privacyContent: string;
  /** @maxLength 200 */
  legalName: string;
  /** @maxLength 1000 */
  businessAddress: string;
  /** @maxLength 320 */
  supportEmail: string;
  /** @maxLength 320 */
  privacyEmail: string;
  /** @maxLength 80 */
  supportPhone: string;
}