# lib/api-zod/src/generated/types/siteContentResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SiteContent } from './siteContent';

export interface SiteContentResponse {
  content: SiteContent;
}